Minor changes (from float to double in Keyword score)

+++ src/main/java/eu/socialsensor/framework/common/domain/Keyword.java
@@ -5,7 +5,6 @@ import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
-import com.google.gson.stream.JsonReader;
 
 /**
  * @author etzoannos - e.tzoannos@atc.gr
@@ -19,13 +18,13 @@ public class Keyword implements JSONable {
 	 
 	@Expose
 	@SerializedName(value = "score")
-	private float score;
+	private double score;
 
 	public Keyword() {        
 	 
 	}
     
-	public Keyword(String keyword, float score) {
+	public Keyword(String keyword, double score) {
 		this.keyword = keyword;
 		this.score = score;
 	}
@@ -38,11 +37,11 @@ public class Keyword implements JSONable {
 		this.keyword = keyword;
 	}
 
-	public float getScore() {
+	public double getScore() {
 		return score;
 	}
 
-	public void setScore(float score) {
+	public void setScore(double score) {
 		this.score = score;
 	}
